fix: move Body to index file to fix build errors
