+++ src/http/body/index.ts
@@ -1,4 +1,14 @@
 import BodyCreator from './body-creator';
-import Body from './body';
+import MediaFormat from '../media-format';
 
-export { BodyCreator, Body };
+export interface Body {
+  length: number;
+  mediaFormat: MediaFormat;
+
+  isEmpty(): boolean;
+  equals(otherBody: Body): boolean;
+  toBuffer(): Buffer;
+  toJson(): any;
+}
+
+export { BodyCreator };
